23 prefetch pages (#24)
* Refactor(login): now in pages instead of components

* Feat(useRooms): added prefetching for room
