+++ hooks/useRooms.ts
@@ -1,4 +1,4 @@
-import { useCallback, useContext } from "react";
+import { useCallback, useContext, useEffect } from "react";
 import { useRouter } from "next/router";
 import { SocketContext } from "../context/socket";
 import { joinRoom as userJoinRoom } from "../api/userApi";
@@ -8,6 +8,10 @@ export const useRooms = () => {
   const router = useRouter();
   const socket = useContext(SocketContext);
   const { changeRoom } = useUser();
+
+  useEffect(() => {
+    router.prefetch("/room");
+  }, [router]);
 
   const joinRoom = useCallback(
     async ({ userId, roomId }: { roomId: string; userId: string }) => {
